Adds random nonce (r) option to `hf mf sim` and `hf 14a sim`.
This makes the PM3 generate random nonces rather than sequential nonces, to
make it act a bit more like a "real" MFC card.  A reader would otherwise be
able to detect the PM3 probing based on the low and sequential nonces and
throw different authentication challenges.

The code includes an implementation of a rand-like function which is seeded
automatically based on the time it takes between the PM3 starting up and the
first call to the RNG.

This isn't cryptographically random, but should be "good enough" to be able
to evade basic detection.